feat: extract gateway_profiles_mcp and mount via capability registry
Move the three public profile MCP tools (get_profile, get_profile_posts,
get_profile_notes) out of gateway_oss.mcp.app into a standalone
gateway_profiles_mcp package registered through the capability registry
entry-point, matching the pattern already used for posts/notes/comments/
following MCP capabilities. get_profile_notes now composes through
NotesService.get_notes_for_profile from gateway_notes instead of the prior
(mismatched) PostsService call.
