+++ packages/gateway_profiles_mcp/pyproject.toml
@@ -1,0 +1,22 @@
+[project]
+name = "gateway_profiles_mcp"
+version = "3.1.0"
+requires-python = ">=3.11"
+dependencies = ["gateway_core", "gateway_mcp_common", "gateway_profiles", "gateway_posts", "gateway_notes", "fastmcp>=3.4.2"]
+
+[project.entry-points."substack_gateway.capabilities"]
+profiles_mcp = "gateway_profiles_mcp:capability"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/gateway_profiles_mcp"]
+
+[tool.uv.sources]
+gateway_core = { workspace = true }
+gateway_mcp_common = { workspace = true }
+gateway_profiles = { workspace = true }
+gateway_posts = { workspace = true }
+gateway_notes = { workspace = true }
