+++ pyproject.toml
@@ -39,6 +39,7 @@ gateway_posts_mcp = { workspace = true }
 gateway_comments_rest = { workspace = true }
 gateway_comments_mcp = { workspace = true }
 gateway_profiles_rest = { workspace = true }
+gateway_profiles_mcp = { workspace = true }
 gateway_rest_common = { workspace = true }
 gateway_mcp_common = { workspace = true }
 
@@ -52,7 +53,7 @@ select = ["E", "F", "I", "UP"]
 ignore = ["E501"]
 
 [tool.ruff.lint.isort]
-known-first-party = ["gateway_core", "gateway_comments", "gateway_following", "gateway_notes", "gateway_notes_rest", "gateway_notes_mcp", "gateway_following_rest", "gateway_following_mcp", "gateway_posts", "gateway_posts_rest", "gateway_posts_mcp", "gateway_comments_rest", "gateway_comments_mcp", "gateway_profiles", "gateway_profiles_rest", "gateway_oss", "gateway_rest_common", "gateway_mcp_common", "substack_gateway"]
+known-first-party = ["gateway_core", "gateway_comments", "gateway_following", "gateway_notes", "gateway_notes_rest", "gateway_notes_mcp", "gateway_following_rest", "gateway_following_mcp", "gateway_posts", "gateway_posts_rest", "gateway_posts_mcp", "gateway_comments_rest", "gateway_comments_mcp", "gateway_profiles", "gateway_profiles_rest", "gateway_profiles_mcp", "gateway_oss", "gateway_rest_common", "gateway_mcp_common", "substack_gateway"]
 
 [tool.ruff.format]
 quote-style = "double"
